+++ unittest/SqlEngineTestBase.cpp
@@ -24,7 +24,7 @@ BOOST_AUTO_TEST_CASE(SQLParserStatementAddTest)
 {
   SQLStatement sqlStmt;
   
-  for (int n=0; n<100; n++) {
+  for (size_t n=0; n<100; n++) {
     SQLSelect *sqlCmd = new SQLSelect();
     sqlStmt.addChildNode(sqlCmd);
     //BOOST_CHECK(sqlStmt.numChildNodes() == (n+1));
